[VersionControl] Simplify the IVersionControl<T> interfaces
We are breaking API/ABI with the newresolver update so we are also taking
this opportunity to implement some improvements in the Version Control
interfaces for blameview/logview etc. The IFastVersionControlViewHandler<T>
interfaces are replacing the current IVersionControlViewHandler<T> interfaces.

This change will also fix bug #4206 as the new view handlers can figure out
if they can actually attach to the current primary view. For example BlameView
requires that the primary view implements ITextFile so now that can be queried
in the 'CanHandle' call.